requirements.txt
# Advanced Clevrr Computer - 100% Local AI Agent
# All dependencies run locally. No external APIs required.
#
# IMPORTANT: Ollama must be installed as a system tool, NOT via pip.
# Install Ollama from https://ollama.ai
# Then run:
#   ollama serve
#   ollama pull llava
#   ollama pull llama3
#
# Ollama provides the local LLM runtime for vision (llava) and
# text reasoning (llama3). It runs as a system service on localhost:11434.

# --- Computer Control ---
pyautogui                          # Mouse + keyboard automation
pywinauto; sys_platform == "win32" # Semantic UI element finding (Windows only)
mss                                # Multi-monitor screen capture
opencv-python                      # Image processing and comparison
uiautomation; sys_platform == "win32"  # Windows UI Automation (Windows only)
pyperclip                          # Clipboard access for text pasting

# --- App Control ---
playwright                         # Browser automation (all platforms)
pytesseract                        # OCR support (optional, for advanced vision)

# --- Local AI (Voice) ---
openai-whisper                     # Offline speech-to-text (runs locally)
pyttsx3                            # Offline text-to-speech (runs locally)
sounddevice                        # Microphone audio recording
soundfile                          # Audio file handling
scipy                              # Signal processing utilities

# --- Local AI (Memory & Search) ---
sentence-transformers              # Local semantic search embeddings

# --- UI & Config ---
gradio                             # Local web dashboard (localhost:7860)
pyyaml                             # YAML config file parsing
Pillow                             # Image handling
numpy                              # Array/numerical processing
psutil                             # System health/process metrics
python-dotenv                      # Environment variable management

# --- HTTP (local only, for Ollama communication) ---
requests                           # HTTP client for local Ollama API
pytest>=7.0.0
pytest-timeout>=2.1.0
